+++ tests/test_device.py
@@ -32,6 +32,25 @@ class DeviceAPIBase(APIBase):
         rtol=1.0e-6,
         atol=0.0,
     ):
+        pytorch_result = str(pytorch_result).replace("cuda", "gpu")
+        if "cpu:" in pytorch_result:
+            pytorch_result = "cpu"
+        assert pytorch_result == paddle_result
+
+
+class PaddleDeviceAPIBase(APIBase):
+    def compare(
+        self,
+        name,
+        pytorch_result,
+        paddle_result,
+        check_value=True,
+        check_shape=True,
+        check_dtype=True,
+        check_stop_gradient=True,
+        rtol=1.0e-6,
+        atol=0.0,
+    ):
         pytorch_result = str(pytorch_result)
         if pytorch_result == "cuda":
             pytorch_result = "cuda:0"
@@ -40,7 +59,7 @@ class DeviceAPIBase(APIBase):
         assert pytorch_result == paddle_result
 
 
-obj = DeviceAPIBase("torch.device")
+obj = PaddleDeviceAPIBase("torch.device")
 
 
 def test_case_1():
