+++ tests/test_get_default_device.py
@@ -14,27 +14,7 @@
 
 import textwrap
 
-from apibase import APIBase
-
-
-class DeviceAPIBase(APIBase):
-    def compare(
-        self,
-        name,
-        pytorch_result,
-        paddle_result,
-        check_value=True,
-        check_shape=True,
-        check_dtype=True,
-        check_stop_gradient=True,
-        rtol=1.0e-6,
-        atol=0.0,
-    ):
-        pytorch_result = str(pytorch_result).replace("cuda", "gpu")
-        if "cpu:" in pytorch_result:
-            pytorch_result = "cpu"
-        assert pytorch_result == paddle_result
-
+from test_device import DeviceAPIBase
 
 obj = DeviceAPIBase("torch.get_default_device")
